CORE-585 - Rearrangements in the way we sent out registration emails.
Untangle event handling for additional state transitions and email sending. Removed the additional state tracking for email sending. Tweaked the send out of emails to solely react on state transitions avoiding the need for additional HTTP resources to explicitly do so.

Extracted TrackedCaseEmails to capture the essence of sending out emails for initial welcome and registration and record the outcome of those activities in the form of comments on the tracked case. We now clearly separate between event listeners that trigger additional state transitions and ones the trigger emails being sent.

Tweaked the names of event listener methods to prepare enhanced documentation via Moduliths, as the Javadoc extraction by Spring Auto RESTDocs cannot extract all documentation of overloaded methods correctly [0].

Introduced CommentFactory to easily create comments based on system activities that can be either triggered automatically or indirectly through a user. In the former case, we use the system's application name as author. Messages are i18ned using the default resource bundle and use a predefined key (see CommentKey) which are appended with .success or .failure to easily create two flavors of the message. This fundamentally implements CORE-642 but needs another round of review to validate the wording for the individual cases.

Changed the CaseStatusUpdated event to only carry the identifier of the case so that event listeners need to look up the case again. This is to make sure that e.g. listeners that work asynchronously or post-commit have to re-lookup the entity.

RegistrationManagement now actively rejects a missing email on attempts to initiate the registration and adds failure and success comments to the tracked case depending on the outcome of the method.

Resolved a package cycle accidentally introduced by the polishing for CORE-615 (3292e7). ActivationCodeProperties is now back in the form of an interface implemented by RegistrationProperties so the the activation code subsystem can relate to it without referring back to the department package.

Removed obsolete Task class.

EmailSender now contains an initialized flag to effectively submit an email to the SMTP server only if that's set to true. The flag is enabled on ApplicationReadyEvent. That was introduced to be able simulate those emails being sent out during demo data creation to see the subsequent domain object modifications executed ("Registration email was sent." et al.

[0] https://github.com/ScaCap/spring-auto-restdocs/issues/434.

Original pull request: #165.
Related tickets: CORE-642.