CORE-613 - Protect endpoint for visitor group submissions with new role.
The HTTP resource to submit visitor groups is now expecting a newly introduced role that does not have any users yet. We do this to be able to merge the branch so that the API to create events and occasion codes can already be used by the frontend team to build UI for that.

Also, third-party integrators can already play with the submission API by tweaking QuaranoWebSecurityConfigurerAdapter as documented in the comment. Running the application locally then allows to submit visitor groups as documented in the API docs without any authentication.