CORE-613 - Enable X.509 authentication for external endpoints.
We now require authentication to the role of THIRD_PARTY for all requests to URIs starting with /ext. This currently contains the API to submit visitor groups for occasions.

To test this, the repository now contains a sample client and server key, key store and trust store infrastructure. The key- and trust stores are also bundled with the artifact if the localDev Maven profile is activated during the build. This allows the certificate authentication to also be used during local executions.

The API root resource now exposes some links that are dependent on the current user's role:

* login, reset-password - if the resource is accessed without authentication at all or any role that's assignable to humans.
* submit-visitors - if the resource is accessed by third-party software

Temporarily introduced a patched copy of Spring HATEOAS' AnnotationMappingDiscoverer to benefit from the fix for [0] until we can upgrade to a released version of it.

Additional security reference documentation for developers, deployments and third-party integrators.

[0] https://github.com/spring-projects/spring-hateoas/issues/1413