CORE-613 - Move all certification authentication config to integration test profile.
Partial configuration in the main application.properties file causes the SSL support to be enabled on startup. As long as the deployments do not have a certificate store configured from the outside yet, this causes the application startup to fail. The cleanup here prevents this but also requires the environments to explicitly configure a few more properties than anticipated before. The deployment documentation was updated accordingly.