CORE-613 - Tweak server setup for local execution without strict need for HTTPS.
We now explicitly configure the develop profile to run the server on port 8443 as it enables SSL support. We also register an addition connector to serve plain HTTP traffic on port 8080.

Simple access to the API is now possible like this again:

$ http :8080

HTTPS based access is supported like this:

$ http --verify=no https://localhost:8443

Note, that we need to disable the verification of the server certificate as it is self-signed and would otherwise be rejected. Also note how we switched to https for the protocol and port.

Access the application and authenticate using a client certificate using additional options:

$ http --verify=no \
 --cert=src/test/resources/security/client/acme-event-manager.crt \
 --certkey=src/test/resources/security/client/acme-event-manager.key \
 https://localhost:8443

The environment specification for the frontend is now turned back to use HTTP.