CORE-294 implements anonymization and partial deletion of personal data
The idea with GDPR implementation is to only overwrite the personal data
(### for text makes anonymization clear or Null for other data types)
and not delete the whole records. By anonymizing, we avoid some problems
with dependencies between data that have to be taken into account in
different places if anything is to be deleted. Besides, the possibility
for statistical analysis e.g. on the level of the postal code or the
city remains.

The following three domains in the data are handled separately as
follows:

1. Once a day during the night, a job runs that deletes all occasions
that were created more than 6 months ago. Furthermore, for all occasions
created more than 16 days ago, the associated visitors are deleted. This
data is so loosely connected to the rest of the data model that deletion
is easy to implement.

2. Once a day during the night, a job runs that anonymizes (for what
exactly see later overview) all contacts (not contact persons or contact
cases) that were created more than 16 days ago. Only from non-anonymized
contacts cases are created, in case of later conversion.

3. Once a day during the night, a job runs that anonymizes all cases
created more than 6 months ago. Thereby all relevant referenced data
(e.g. account, person, ...) are anonymized (what exactly see later
overview). Anonymized cases are no longer included in the overviews and
exports. Therefore the cases get a corresponding status.

When to delete is adjustable.

To enable tests of these time-related functions, it is now possible to
specify a delta for the creation and modification time of the data via
the property quarano.date-time.delta. This only works in non-production
environments.