Correction of fixed-sender address of cuxhaven stage.
It seems that there is a problem if sender and recipient address are the same, because mails have not been delivered in that case to the original owner of this mailbox. As a fix we created to addresses for cuxhaven stage that fixed-sender and fixed-recipient distinguish.

(cherry picked from commit f87ed9447253a2e513523e34c845e3c150278f98)